add and verify BankAccount changed, send deleted
