--- client/src/main/kotlin/com/r3/corda/finance/cash/issuer/client/flows/ReceiveVerifyBankAccount.kt
@@ -1,20 +1,0 @@
-package com.r3.corda.finance.cash.issuer.client.flows
-
-import co.paralleluniverse.fibers.Suspendable
-import com.r3.corda.finance.cash.issuer.common.flows.AbstractVerifyBankAccount
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.FlowSession
-import net.corda.core.flows.InitiatedBy
-import net.corda.core.flows.ReceiveFinalityFlow
-
-@InitiatedBy(AbstractVerifyBankAccount::class)
-class ReceiveVerifyBankAccount(val otherSession: FlowSession) : FlowLogic<Unit>() {
-    @Suspendable
-    override fun call() {
-        logger.info("Starting ReceiveVerifyBankAccount flow...")
-        //return subFlow(ReceiveTransactionFlow(otherSession, true, StatesToRecord.ALL_VISIBLE))
-        if (!serviceHub.myInfo.isLegalIdentity(otherSession.counterparty)) {
-            subFlow(ReceiveFinalityFlow(otherSession))
-        }
-    }
-}
