--- common/src/main/kotlin/com/r3/corda/finance/cash/issuer/common/flows/AbstractSendBankAccount.kt
@@ -1,9 +1,0 @@
-package com.r3.corda.finance.cash.issuer.common.flows
-
-import com.r3.corda.finance.cash.issuer.common.states.BankAccountState
-import net.corda.core.contracts.StateAndRef
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.InitiatingFlow
-
-@InitiatingFlow
-abstract class AbstractSendBankAccount : FlowLogic<StateAndRef<BankAccountState>>()
