--- common/src/main/kotlin/com/r3/corda/finance/cash/issuer/common/flows/AbstractVerifyBankAccount.kt
@@ -1,10 +1,0 @@
-package com.r3.corda.finance.cash.issuer.common.flows
-
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.InitiatingFlow
-import net.corda.core.transactions.SignedTransaction
-
-@InitiatingFlow
-abstract class AbstractVerifyBankAccount : FlowLogic<SignedTransaction>() {
-
-}
