--- service/src/main/kotlin/com/r3/corda/finance/cash/issuer/service/flows/ReceiveBankAccount.kt
@@ -1,19 +1,0 @@
-package com.r3.corda.finance.cash.issuer.service.flows
-
-import co.paralleluniverse.fibers.Suspendable
-import com.r3.corda.finance.cash.issuer.common.flows.AbstractSendBankAccount
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.FlowSession
-import net.corda.core.flows.InitiatedBy
-import net.corda.core.flows.ReceiveTransactionFlow
-import net.corda.core.node.StatesToRecord
-import net.corda.core.transactions.SignedTransaction
-
-@InitiatedBy(AbstractSendBankAccount::class)
-class ReceiveBankAccount(val otherSession: FlowSession) : FlowLogic<SignedTransaction>() {
-    @Suspendable
-    override fun call(): SignedTransaction {
-        logger.info("Starting ReceiveBankAccount flow...")
-        return subFlow(ReceiveTransactionFlow(otherSession, true, StatesToRecord.ALL_VISIBLE))
-    }
-}
