+++ service/src/main/kotlin/com/r3/corda/finance/cash/issuer/service/flows/VerifyBankAccount.kt
@@ -2,11 +2,14 @@ package com.r3.corda.finance.cash.issuer.service.flows
 
 import co.paralleluniverse.fibers.Suspendable
 import com.r3.corda.finance.cash.issuer.common.contracts.BankAccountContract
-import com.r3.corda.finance.cash.issuer.common.flows.AbstractVerifyBankAccount
-import com.r3.corda.finance.cash.issuer.common.types.AccountNumber
-import com.r3.corda.finance.cash.issuer.common.utilities.getBankAccountStateByAccountNumber
+import com.r3.corda.finance.cash.issuer.common.flows.VerifyBankAccountFlow
+import com.r3.corda.finance.cash.issuer.common.utilities.getBankAccountStateByLinearId
 import net.corda.core.contracts.Command
-import net.corda.core.flows.*
+import net.corda.core.contracts.UniqueIdentifier
+import net.corda.core.flows.FinalityFlow
+import net.corda.core.flows.FlowException
+import net.corda.core.flows.StartableByRPC
+import net.corda.core.flows.StartableByService
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionBuilder
 
@@ -16,19 +19,20 @@ import net.corda.core.transactions.TransactionBuilder
  */
 @StartableByService
 @StartableByRPC
-class VerifyBankAccount(val accountNumber: AccountNumber) : AbstractVerifyBankAccount() {
+class VerifyBankAccount(val linearId: UniqueIdentifier) : VerifyBankAccountFlow.AbstractVerifyBankAccount() {
 
     @Suspendable
     override fun call(): SignedTransaction {
         logger.info("Starting VerifyBankAccount flow.")
 
         val notary = serviceHub.networkMapCache.notaryIdentities.first()
-        val bankAccountStateAndRef = getBankAccountStateByAccountNumber(accountNumber, serviceHub)
-                ?: throw FlowException("Bank account $accountNumber not found.")
+        val bankAccountStateAndRef = getBankAccountStateByLinearId(linearId, serviceHub)
+                ?: throw FlowException("Bank account by linearId $linearId not found.")
 
+        //TODO move validation into BankAccountContract.verify
         val bankAccountState = bankAccountStateAndRef.state.data
         if (bankAccountState.verified) {
-            throw FlowException("Bank account $accountNumber is already verified.")
+            throw FlowException("Bank account ${bankAccountState.accountNumber} is already verified.")
         }
 
         val ownerSsession = initiateFlow(bankAccountState.owner)
@@ -40,7 +44,9 @@ class VerifyBankAccount(val accountNumber: AccountNumber) : AbstractVerifyBankAc
                 .addInputState(bankAccountStateAndRef)
                 .addCommand(command)
                 .addOutputState(updatedBankAccountState, BankAccountContract.CONTRACT_ID)
+
         val stx = serviceHub.signInitialTransaction(utx)
+
         // Share the updated bank account state with the owner.
         val sessionsForFinality = if (serviceHub.myInfo.isLegalIdentity(bankAccountState.owner)) emptyList() else listOf(ownerSsession)
         return subFlow(FinalityFlow(stx, sessionsForFinality))
